♻ Extract new component JobHeader from Chat for the Job process
